Enable setting which component is used to render attachments
